Standalone 261

+++ Template/LO/SubProcesses/refine.sh
@@ -5,17 +5,15 @@ if [ $CLUSTER_LHAPATH ]; then
   export LHAPATH=$CLUSTER_LHAPATH;
 fi
 
-# Add CVMFS libraries to LD_LIBRARY_PATH, if not present already
-if [ $SRT_LD_LIBRARY_PATH_SCRAMRT ]; then 
-  if [ -n ${LD_LIBRARY_PATH##*${SRT_LD_LIBRARY_PATH_SCRAMRT}*} ]; then
-    export LD_LIBRARY_PATH="$SRT_LD_LIBRARY_PATH_SCRAMRT:$LD_LIBRARY_PATH"
-  fi
-fi
-
 if [[ -e MadLoop5_resources.tar.gz && ! -e MadLoop5_resources ]]; then
 tar -xzf MadLoop5_resources.tar.gz
 fi
-k=%(name)s_app.log
+keeplog=%(keeplog)s
+if [ "$keeplog" = true ] ; then
+    k=%(name)s_app.log
+else
+    k=/dev/null
+fi
 script=%(script_name)s                         
 
 grid_directory=%(base_directory)s
@@ -31,9 +29,10 @@ j=%(directory)s
           	 fi
           fi	
      fi
-     
      cd $j
-     rm -f $k
+     if [ "$keeplog" = true ] ; then
+	 rm -f $k
+     fi
      rm -f moffset.dat >& /dev/null
       echo   %(offset)s  > moffset.dat
      if  [[ -e ftn26 ]]; then
@@ -57,17 +56,27 @@ j=%(directory)s
      # filesystem problem (executable not found)
      for((try=1;try<=16;try+=1)); 
      do
-         ../madevent 2>&1 >> $k <input_sg.txt | tee -a $k;
-     status_code=${PIPESTATUS[0]};
-         if [ -s $k ]
-         then
-             break
-         else
-             echo $try > fail.log 
-         fi
+	 if [ "$keeplog" = true ] ; then
+             %(Ppath)s/madevent 2>&1 >> $k <input_sg.txt | tee -a $k;
+	     status_code=${PIPESTATUS[0]};
+             if [ -s $k ]
+             then
+		 break
+             else
+		 echo $try > fail.log 
+             fi
+	 else
+	     %(Ppath)s/madevent 2>&1 >> log.txt <input_sg.txt | tee -a log.txt;
+	     status_code=${PIPESTATUS[0]};
+	     if [ -s log.txt ]
+             then
+		 rm log.txt
+                 break
+             else
+                  echo $try > fail.log
+             fi
+	 fi
      done
-     echo "" >> $k; echo "ls status:" >> $k; ls >> $k
-     cat $k >> log.txt
      if [[ $status_code -ne 0 ]]; then 
 	 rm results.dat
 	 echo "ERROR DETECTED"
@@ -76,5 +85,16 @@ j=%(directory)s
      if [[ -e ftn26 ]]; then
          cp ftn26 ftn25
      fi
+
+     if [ "$keeplog" = true ] ; then
+	 echo "" >> $k; echo "ls status:" >> $k; ls >> $k	 
+     else
+	 rm ftn26 &> /dev/null
+     fi
+
+
+
+
+
      cd ../
